requirements.txt
# Web framework
streamlit>=1.29.0

# PyTorch with CUDA support
# Note: For CUDA 13, you may need to install PyTorch from the official index:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu131
# Or use the latest version that supports your CUDA version
# torch>=2.1.0
torchaudio>=2.1.0

# Transformers and related libraries
transformers>=4.38.0
datasets>=2.16.0
sentencepiece>=0.1.96

# Russian accent/stress marking
ruaccent>=1.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# Utilities
numpy>=1.24.0
scipy>=1.10.0

omegaconf>=2.3.0